# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Twitee API",
        "version": "1.0.0"
    },
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "description": "Provide email and password to register user",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Provide user details for registration",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "user"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@email.com"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-06-02T19:08:17.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-06-02T19:08:17.000000Z"
                                        },
                                        "id": {
                                            "type": "string",
                                            "format": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fields are required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "fields are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate user",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "email required, password is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout authenticated user",
                "description": "Logout user",
                "operationId": "authLogout",
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/user-profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Show user profile",
                "description": "Display user profile",
                "operationId": "displayUserProfile",
                "responses": {
                    "200": {
                        "description": "User Profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "ade"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "user@email.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/comment/{post}": {
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Add a comment",
                "description": "Add comment to a post",
                "operationId": "addComment",
                "requestBody": {
                    "description": "Add a comment",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "comment_text"
                                ],
                                "properties": {
                                    "comment_text": {
                                        "type": "string",
                                        "example": "This is a comment text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comment added successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "post_id": {
                                            "type": "integer",
                                            "example": "4"
                                        },
                                        "comment_text": {
                                            "type": "string",
                                            "example": "test comment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fields are required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "fields are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "comment/{post_id}/{id}": {
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Delete comment",
                "description": "Delete post comment",
                "operationId": "deleteComment",
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Post or comment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post or comment does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You don't have access to delete this comment!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "allPosts",
                "responses": {
                    "200": {
                        "description": "All Posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Jon Write"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "The goldie is back"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create a post",
                "description": "User create a post",
                "operationId": "createPost",
                "requestBody": {
                    "description": "Provide post details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "The goldie is back"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post created successfully!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Jon Write"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "The goldie is back"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Fields are required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "fields are required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/post/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get a post detail",
                "description": "Returns detail of a particular post",
                "operationId": "getPostDetail",
                "responses": {
                    "200": {
                        "description": "Post detail:",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "Jon Write"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "The goldie is back"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "post does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete a post",
                "description": "Delete a post along with it's comments",
                "operationId": "deletePost",
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You don't have access to delete this post: The goldie is back!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}